% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBreaker.R
\name{plot.resTreeBreaker}
\alias{plot.resTreeBreaker}
\title{Plotting function}
\usage{
\method{plot}{resTreeBreaker}(res, type = "cons")
}
\arguments{
\item{res}{An object of class resTreeBreaker}

\item{type}{Type of plot to perform, can be cons (default), states, traces or correlation}
}
\value{
Invisible object
}
\description{
Plotting function
}
