% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBreaker.R
\name{readOutFile}
\alias{readOutFile}
\title{Reading an output file}
\usage{
readOutFile(outfile)
}
\arguments{
\item{outfile}{name of the output file}
}
\value{
object containing output
}
\description{
Reading an output file
}
