% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBreaker.R
\name{treeBreaker}
\alias{treeBreaker}
\title{treeBreaker main function}
\usage{
treeBreaker(inputfile_tree, inputfile_phenotype, outputfile, x = 5e+05,
  y = 5e+05, z = 1000, seed = NA, verbose = F)
}
\arguments{
\item{inputfile_tree}{newick file to use as input tree}

\item{inputfile_phenotype}{tab file to use as input phenotype}

\item{outputfile}{output file}

\item{x}{the number of iterations after burn-in (default is 500000)}

\item{y}{the number of burn-in iterations (default is 500000)}

\item{z}{the number of iterations between samples (default is 1000)}

\item{seed}{the seed for the random number generator (default is a random seed)}

\item{verbose}{verbose mode}
}
\value{
success status
}
\description{
treeBreaker main function
}
